create base class for SR2ML models
