+++ src/Bases/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2020, Battelle Energy Alliance, LLC
+# ALL RIGHTS RESERVED
+
+from .ModelBase import ModelBase
