+++ src/Reliabilities/LogLinearModel.py
@@ -5,14 +5,7 @@ Created on Feb. 7, 2020
 
 @author: wangc, mandd
 """
-#for future compatibility with Python 3--------------------------------------------------------------
-from __future__ import division, print_function, unicode_literals, absolute_import
-#End compatibility block for Python 3----------------------------------------------------------------
-
 #External Modules------------------------------------------------------------------------------------
-import abc
-import sys
-import os
 import numpy as np
 import numpy.ma as ma
 from scipy.integrate import quad
@@ -56,20 +49,20 @@ class LogLinearModel(ReliabilityBase):
       @ In, None
       @ Out, None
     """
-    ReliabilityBase.__init__(self)
+    super().__init__()
     # Shape parameter
     self._alpha = 1.
     # The inverse is the scale parameter
     self._beta = 1.
 
-  def _localHandleInput(self, paramInput):
+  def _handleInput(self, paramInput):
     """
       Function to read the portion of the parsed xml input that belongs to this specialized class
       and initialize some stuff based on the inputs got
       @ In, paramInput, InputData.ParameterInput, the parsed xml input
       @ Out, None
     """
-    ReliabilityBase._localHandleInput(self, paramInput)
+    super()._handleInput(paramInput)
     for child in paramInput.subparts:
       if child.getName().lower() == 'alpha':
         self._alpha = self.setVariable(child.value)
@@ -87,7 +80,7 @@ class LogLinearModel(ReliabilityBase):
       @ In, inputDict, dict, dictionary of inputs
       @ Out, None
     """
-    ReliabilityBase.initialize(self, inputDict)
+    super().initialize(inputDict)
 
   def _probabilityFunction(self):
     """
