+++ src/Reliabilities/__init__.py
@@ -5,8 +5,6 @@ Created on Feb. 4, 2020
 @author: wangc, mandd
 """
 
-from __future__ import absolute_import
-
 from .ExponentialModel import ExponentialModel
 from .ErlangianModel import ErlangianModel
 from .GammaModel import GammaModel
